Harmony
Project name is Harmony.  It is a web application based international conference held by CHRIST (Deemed to Be University). Where dashboard is created for user and admin. User can view his profile, upload paper and see the status of the paper. Where Admin can see people register for the conference as well as registered for Accommodation. Admin can download paper of the author and approve them. The technology used in this application are bootstrap, HTML,CSS, Java Script and jsp. Link of the conference website: -www.harmony2019.in